Move worker test files to folder
