+++ client/src/worker/tests/util.ts
@@ -1,6 +1,6 @@
 import path from "path";
 import fs from "fs";
-import Go from "../worker/go";
+import Go from "../go";
 
 export interface TestFile {
   name: string;
@@ -20,7 +20,7 @@ export function newTestFile(name: string, size = 1024): TestFile {
 
 export async function initGo() {
   const go = new Go();
-  const wasmPath = path.join(__dirname, "../../dist/wormhole.wasm");
+  const wasmPath = path.join(__dirname, "../../../dist/wormhole.wasm");
   await WebAssembly.instantiate(
     fs.readFileSync(wasmPath),
     go.importObject
